Nextjs: add loading stream demo
