+++ src/app/(demo)/routing/loading.tsx
@@ -1,0 +1,5 @@
+const LoadingInDemo = () => {
+  return <div>LoadingInDemo</div>;
+};
+
+export default LoadingInDemo;
