+++ src/app/(demo)/routing/template.tsx
@@ -1,9 +1,10 @@
 import Link from 'next/link';
 import {
+  ROUTE_DEMO_LOADING_STREAM,
   ROUTE_DEMO_ROUTING_CATCH_ALL_SEGMENTS,
   ROUTE_DEMO_ROUTING_OPTIONAL_CATCH_ALL_SEGMENTS,
 } from '@routes/pathes';
-import Header from '@component/header';
+import Header from '@component/general/header';
 
 interface IRoutingTemplateProps {
   children: React.ReactNode;
@@ -13,7 +14,6 @@ const RoutingTemplate: React.FC<IRoutingTemplateProps> = (props) => {
   const { children } = props;
   return (
     <div>
-      {children}
       <Header>
         <Link href={ROUTE_DEMO_ROUTING_CATCH_ALL_SEGMENTS}>
           Catch-all segements
@@ -21,7 +21,9 @@ const RoutingTemplate: React.FC<IRoutingTemplateProps> = (props) => {
         <Link href={ROUTE_DEMO_ROUTING_OPTIONAL_CATCH_ALL_SEGMENTS}>
           Optional catch-all segements
         </Link>
+        <Link href={ROUTE_DEMO_LOADING_STREAM}>Loading & steaming</Link>
       </Header>
+      {children}
     </div>
   );
 };
