+++ src/app/(pages)/to-do-list/page.tsx
@@ -1,5 +1,3 @@
-import Template from '@app/template';
-
 const TodoList = () => {
   return <div>TodoList</div>;
 };
