+++ src/app/loading.tsx
@@ -1,7 +1,9 @@
 'use client';
 
-const Loading = () => {
-  return <div>Loading</div>;
+interface ILoadingProps {}
+const Loading: React.FC<ILoadingProps> = (props) => {
+  return <div>Loading </div>;
 };
 
+Loading.defaultProps = {};
 export default Loading;
