+++ src/components/general/header/header.module.scss
@@ -1,0 +1,8 @@
+.container {
+  max-width: var(--max-width);
+  margin: 0 auto;
+  display: flex;
+  justify-content: center;
+  column-gap: 32px;
+  padding: var(--large-padding);
+}
