--- src/components/header/header.module.scss
@@ -1,8 +1,0 @@
-.container {
-  max-width: var(--max-width);
-  margin: 0 auto;
-  display: flex;
-  justify-content: center;
-  column-gap: 32px;
-  padding: var(--large-padding);
-}
