+++ src/components/post/post.tsx
@@ -1,0 +1,21 @@
+const Post = () => {
+  return (
+    <div>
+      Post
+      <video width='320' height='240' controls>
+        <source src='/Black-Friday.mp4' />
+      </video>
+      <video width='320' height='240' controls>
+        <source src='/Black-Friday.mp4' />
+      </video>
+      <video width='320' height='240' controls>
+        <source src='/Black-Friday.mp4' />
+      </video>
+      <video width='320' height='240' controls>
+        <source src='/Black-Friday.mp4' />
+      </video>
+    </div>
+  );
+};
+
+export default Post;
